Collision algorithm, smaller type changes.

+++ bubbleSim/simulation.h
@@ -51,7 +51,7 @@ class Simulation {
     }
   };
 
-  int getStep() { return m_step; }
+  unsigned int getStep() { return m_step; }
 
   void step(PhaseBubble& bubble, numType t_dP);
   /*
@@ -104,7 +104,7 @@ class Simulation {
   // Cumulative time
   int m_seed;
   numType m_time = 0.;
-  size_t m_step = 0;
+  u_int m_step = 0;
   // One step time length
   numType m_dt;
   numType m_step_dt;
